+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 - Add path parameter sanitization
+- Add support for the classification field for Files and Folders
 - Fix bug with notification email field deserializing for `BoxUser` 
 
 ## 3.23.0 [2020-05-12]